-c1-production -c2-producer -t12n-producing -mmmi- 2013-12-11 15-31-25 app:SVN7151M, app-cidadedecuritiba:SVN238, app-core:SVN2024, app-frontpage:SVN215, app-sysutils:SVN147, app-veriwell:SVN1107, nodeapp:SVN1496 -  at usloft2405 at ServerLoft - serverloft.com - Fujistu 2 processors Intel(R) Xeon(R) E5520 @ 2.27GHz Processors - 16 cores - 12.0GB RAM 1TB RAID1 HDD storage 60GB RAID1 SSD storage high-speed Internet Windows Server 2008 R2 Standard 64-bit Service Pack 1 at Saint Louis, Missouri - United States, North America, this machine and telmo to speed up effectiveness, performance and build funcionality and readiness!!!!!!!!#define THIS_PRODUCT_VERSION "2013-12-11 15-31-25 app:SVN7151M, app-cidadedecuritiba:SVN238, app-core:SVN2024, app-frontpage:SVN215, app-sysutils:SVN147, app-veriwell:SVN1107, nodeapp:SVN1496\0" #define THIS_FILE_VERSION "2013-12-11 15-31-25 app:SVN7151M, app-cidadedecuritiba:SVN238, app-core:SVN2024, app-frontpage:SVN215, app-sysutils:SVN147, app-veriwell:SVN1107, nodeapp:SVN1496\0" #define __THIS_PRODUCT_VERSION 2013,1211,1531,25 #define __THIS_FILE_VERSION 2013,1211,1531,25

+++ appseed/base/base/hotplugin/hotplugin_host.cpp
@@ -565,7 +565,7 @@ throw todo(get_thread_app());
 
       m_dib->map();
 
-      memcpy(m_dib->m_pcolorref, m_pcolorref, m_dib->area() * sizeof(COLORREF));
+      memcpy(m_dib->m_pcolorref, m_pcolorref, (size_t) (m_dib->area() * sizeof(COLORREF)));
 
       pgraphics->BitBlt(lprect->left, lprect->top, m_sizeBitmap.cx, m_sizeBitmap.cy, m_dib->get_graphics());
